typename change

+++ include/Random.h
@@ -34,7 +34,7 @@ private:
   boost::mt19937 rnd_;
   double rnd_min_, rnd_width_;
 
-  unsigned int find(const double x, const std::vector<double>& v) const;
+  size_t find(const double x, const std::vector<double>& v) const;
   inline double rand()
   {
     return (rnd_()-rnd_min_)/rnd_width_;
